Break ops tests into multiple files
`test_ops.py` is getting unwieldy and difficult to work with, especially in cases where there are multiple separate tests per op (which may be in different places in the file.) This extract the tests for each op with multiple tests into a separate file for ease of use.
